Simple test correctly working
For the moment, we are using a static experiment “_tester_” to run
tests.

In the future, we gonna have test.php?experiment_id=#id to run tests
using the trained model in #id experiment.

It is not so robust right now, but at least we can use it also from
smartphones